requirements-dev.txt
# Development dependencies
pytest>=7.4.0
pytest-mock>=3.11.0
black>=23.7.0
ruff>=0.0.285
boto3>=1.28.0
botocore>=1.31.0
PyYAML>=6.0
